requirements.txt
# 后端框架
Flask==2.3.3

# 核心图像处理与科学计算
opencv-python==4.8.0.76
Pillow==10.0.1
numpy==1.26.0
matplotlib==3.7.2

# AI与深度学习 (YOLOv8及其依赖)
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.203